Fixed and add create review in review use cases

+++ parka/lib/data/dtos/review/create_review_dto.dart
@@ -2,7 +2,7 @@ class CreateReviewDto {
   String parkingId;
   String reservationId;
   String review;
-  double calification;
+  int calification;
   bool type;
   String title;
